pyproject.toml
[project]
name = "yolo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "numpy>=2.2.6",
    "onnxruntime >= 1.8",
    "pillow>=11.3.0",
    "torchvision>=0.17.2"
]
